basic scene manager

+++ AGame/zChunk.cpp
@@ -7,11 +7,11 @@ Chunk::Chunk(Archetype* holder, const uint32_t& size)
 	_owning_archetype(holder),
 	_number_of_entities(0)
 {
-	int bytesize = 0;
+	size_t bytesize = 0;
 	for (auto i : _owning_archetype->_descriptions) {
 		bytesize += i->_size;
 	}
-	_data = std::make_unique<char[]>(bytesize*size);
+	_data = std::make_unique<char[]>(bytesize*(size_t)size);
 }
 
 int Chunk::Add() {
